create new class instead of new method so that the behaviour can be turned on / off cleanly in production code
