+++ src/test/java/com/alphawallet/attestation/cheque/TestCheque.java
@@ -7,6 +7,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import com.alphawallet.attestation.core.AttestationCrypto;
 import com.alphawallet.attestation.IdentifierAttestation.AttestationType;
+import com.alphawallet.attestation.core.AttestationCryptoWithEthereumCharacteristics;
 import com.alphawallet.attestation.core.SignatureUtility;
 import com.alphawallet.attestation.core.URLUtility;
 import java.io.IOException;
@@ -30,8 +31,8 @@ public class TestCheque {
   public static void setupKeys() throws Exception {
     rand = SecureRandom.getInstance("SHA1PRNG");
     rand.setSeed("seed".getBytes());
-    AttestationCrypto crypto = new AttestationCrypto(rand);
-    senderKeys = crypto.constructECKeysWithLowestYCoord();
+    AttestationCrypto crypto = new AttestationCryptoWithEthereumCharacteristics(rand);
+    senderKeys = crypto.constructECKeys();
   }
